huge refactor; missing: filter integration, adapter, high-level interface
