+++ encryption/publisher/doc.go
@@ -1,0 +1,6 @@
+// Publisher periodically publishes an info about itself on a known channel.
+// This channel is a particular topic calculated from the public key.
+// It is required for other peers to start a secure conversation immediately
+// using distibuted data through the channel.
+
+package publisher
